+++ lib/appsprint_flutter.dart
@@ -1,5 +1,6 @@
 library appsprint_flutter;
 
 export 'src/appsprint.dart';
+export 'src/appsprint_apple_ads.dart';
 export 'src/appsprint_native.dart';
 export 'src/types.dart';